[expo-router] rename Stack.Screen.Title into Stack.Title
