+++ docs/pages/router/advanced/native-tabs.mdx
@@ -802,7 +802,7 @@ import { Stack } from 'expo-router';
 export default function SearchIndex() {
   return (
     <>
-      <Stack.Screen.Title>Search</Stack.Screen.Title>
+      <Stack.Title>Search</Stack.Title>
       <Stack.SearchBar placement="automatic" placeholder="Search" onChangeText={() => {}} />
       <ScrollView>{/* Screen content */}</ScrollView>
     </>
