+++ docs/pages/router/advanced/stack-toolbar.mdx
@@ -363,7 +363,7 @@ export default function RootLayout() {
 
 <Collapsible summary="Large title does not collapse when scrolling">
 
-When using `headerLargeTitle: true` (or `<Stack.Screen.Title large>`) alongside `Stack.Toolbar`, the large title may not collapse on scroll. This happens when the scrollable view is not the direct first child of the screen component.
+When using `headerLargeTitle: true` (or `<Stack.Title large>`) alongside `Stack.Toolbar`, the large title may not collapse on scroll. This happens when the scrollable view is not the direct first child of the screen component.
 
 To fix this, ensure `ScrollView` or `FlatList` is the first child rendered by your screen component. If you need a wrapper, set `collapsable={false}` on it:
 
@@ -375,7 +375,7 @@ import { ScrollView, View, Text } from 'react-native';
 export default function Home() {
   return (
     <ScrollView>
-      <Stack.Screen.Title large>Home</Stack.Screen.Title>
+      <Stack.Title large>Home</Stack.Title>
       <Text>Content here</Text>
     </ScrollView>
   );
@@ -395,7 +395,7 @@ export default function Home() {
     <View collapsable={false}>
       /* @end */
       <ScrollView>
-        <Stack.Screen.Title large>Home</Stack.Screen.Title>
+        <Stack.Title large>Home</Stack.Title>
         <Text>Content here</Text>
       </ScrollView>
     </View>
