+++ packages/expo-router/src/layouts/Stack.web.tsx
@@ -1,10 +1,11 @@
 // This will be changed to `ExperimentalModalStack` in @expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
 // When the `EXPO_UNSTABLE_WEB_MODAL` env variable is truthy.
 import Stack from './_web-modal';
-import { StackScreen, StackHeader, StackToolbar } from './stack-utils';
+import { StackScreen, StackHeader, StackTitle, StackToolbar } from './stack-utils';
 
 Stack.Screen = StackScreen;
 Stack.Header = StackHeader;
+Stack.Title = StackTitle;
 Stack.Toolbar = StackToolbar;
 export { Stack };
