+++ packages/expo-router/src/layouts/stack-utils/StackScreen.tsx
@@ -2,12 +2,8 @@
 import { Children, isValidElement, useMemo, type PropsWithChildren } from 'react';
 
 import { StackHeaderComponent, appendStackHeaderPropsToOptions } from './StackHeaderComponent';
-import {
-  StackScreenTitle,
-  appendStackScreenTitlePropsToOptions,
-  StackScreenBackButton,
-  appendStackScreenBackButtonPropsToOptions,
-} from './screen';
+import { StackTitle, appendStackTitlePropsToOptions } from './StackTitle';
+import { StackScreenBackButton, appendStackScreenBackButtonPropsToOptions } from './screen';
 import { StackToolbar, appendStackToolbarPropsToOptions } from './toolbar';
 import type { ParamListBase, StackNavigationState } from '../../react-navigation/native';
 import type {
@@ -97,23 +93,6 @@ export interface StackScreenProps extends PropsWithChildren {
  *  );
  * }
  * ```
- *
- * @example
- * ```tsx app/home.tsx
- * import { Stack } from 'expo-router';
- *
- * export default function HomePage() {
- *   return (
- *     <>
- *       <Stack.Screen
- *         options={{ headerTransparent: true }}
- *       />
- *       <Stack.Screen.Title>Welcome Home</Stack.Screen.Title>
- *       // Page content
- *     </>
- *   );
- * }
- * ```
  */
 export const StackScreen = Object.assign(
   function StackScreen({ children, options, ...rest }: StackScreenProps) {
@@ -137,7 +116,10 @@ export const StackScreen = Object.assign(
     );
   },
   {
-    Title: StackScreenTitle,
+    /**
+     * @deprecated Use `Stack.Title` instead.
+     */
+    Title: StackTitle,
     BackButton: StackScreenBackButton,
   }
 );
@@ -195,8 +177,8 @@ export function appendScreenStackPropsToOptions(
       return appendStackHeaderPropsToOptions(opts, child.props);
     }
 
-    if (isChildOfType(child, StackScreenTitle)) {
-      return appendStackScreenTitlePropsToOptions(opts, child.props);
+    if (isChildOfType(child, StackTitle)) {
+      return appendStackTitlePropsToOptions(opts, child.props);
     }
 
     if (isChildOfType(child, StackScreenBackButton)) {
