+++ packages/expo-router/src/layouts/stack-utils/index.tsx
@@ -7,6 +7,8 @@ export { StackSearchBar, type StackSearchBarProps } from './StackSearchBar';
 export * from './toolbar';
 export * from './screen';
 
+export { StackTitle, appendStackTitlePropsToOptions, type StackTitleProps } from './StackTitle';
+
 export {
   StackScreen,
   appendScreenStackPropsToOptions,
